requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
twilio==9.3.3
schedule==1.2.2
python-dotenv==1.0.1
